Separate prometheus into different folder structures.
