+++ cmd/start.go
@@ -4,18 +4,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net"
-	"net/http"
-	"os"
-	"strconv"
-	"sync"
-	"time"
-
 	managerpb "github.com/dsrvlabs/vatz-proto/manager/v1"
 	pluginpb "github.com/dsrvlabs/vatz-proto/plugin/v1"
+	"github.com/dsrvlabs/vatz/monitoring/prometheus"
 	"github.com/dsrvlabs/vatz/rpc"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
+	"github.com/dsrvlabs/vatz/utils"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
@@ -23,7 +16,10 @@ import (
 	grpchealth "google.golang.org/grpc/health"
 	healthpb "google.golang.org/grpc/health/grpc_health_v1"
 	"google.golang.org/grpc/reflection"
-	"google.golang.org/protobuf/types/known/emptypb"
+	"net"
+	"os"
+	"strconv"
+	"time"
 
 	"github.com/dsrvlabs/vatz/manager/api"
 	config "github.com/dsrvlabs/vatz/manager/config"
@@ -110,9 +106,9 @@ func initiateServer(ch <-chan os.Signal) error {
 
 	if cfg.Vatz.MonitoringInfo.Prometheus.Enabled {
 		if defaultPromPort == promPort {
-			initMetricsServer(cfg.Vatz.MonitoringInfo.Prometheus.Address, strconv.Itoa(cfg.Vatz.MonitoringInfo.Prometheus.Port), cfg.Vatz.ProtocolIdentifier)
+			prometheus.InitMetricsServer(cfg.Vatz.MonitoringInfo.Prometheus.Address, strconv.Itoa(cfg.Vatz.MonitoringInfo.Prometheus.Port), cfg.Vatz.ProtocolIdentifier)
 		} else {
-			initMetricsServer(cfg.Vatz.MonitoringInfo.Prometheus.Address, promPort, cfg.Vatz.ProtocolIdentifier)
+			prometheus.InitMetricsServer(cfg.Vatz.MonitoringInfo.Prometheus.Address, promPort, cfg.Vatz.ProtocolIdentifier)
 		}
 	}
 
@@ -128,37 +124,11 @@ func initiateServer(ch <-chan os.Signal) error {
 func startExecutor(pluginInfo config.PluginInfo, quit <-chan os.Signal) {
 	//TODO:: value in map would be overridden by different plugins flag value if function name is the same
 	isOkayToSend := false
-	grpcClients := getClients(pluginInfo.Plugins)
+	grpcClients := utils.GetClients(pluginInfo.Plugins)
 	//TODO: Need updated with better way for Dynamic handlers
 	for idx, singleClient := range grpcClients {
 		go multiPluginExecutor(pluginInfo.Plugins[idx], singleClient, isOkayToSend, quit)
 	}
-}
-
-func getClients(plugins []config.Plugin) []pluginpb.PluginClient {
-	var grpcClients []pluginpb.PluginClient
-
-	if len(plugins) > 0 {
-		for _, plugin := range plugins {
-			conn, err := grpc.Dial(fmt.Sprintf("%s:%d", plugin.Address, plugin.Port), grpc.WithInsecure())
-			if err != nil {
-				log.Fatal().Str("module", "main").Msgf("gRPC Dial Error(%s): %s", plugin.Name, err)
-			}
-			grpcClients = append(grpcClients, pluginpb.NewPluginClient(conn))
-		}
-	} else {
-		// TODO: Is this really neccessary???
-		defaultConnectedTarget := "localhost:9091"
-		conn, err := grpc.Dial(defaultConnectedTarget, grpc.WithInsecure())
-		if err != nil {
-			log.Fatal().Str("module", "main").Msgf("gRPC Dial Error: %s", err)
-		}
-
-		//TODO: Please, Create a better client functions with static
-		grpcClients = append(grpcClients, pluginpb.NewPluginClient(conn))
-	}
-
-	return grpcClients
 }
 
 func multiPluginExecutor(plugin config.Plugin,
@@ -197,102 +167,4 @@ func initHealthServer(s *grpc.Server) {
 	gRPCHealthServer := grpchealth.NewServer()
 	gRPCHealthServer.SetServingStatus("vatz-health-status", healthpb.HealthCheckResponse_SERVING)
 	healthpb.RegisterHealthServer(s, gRPCHealthServer)
-}
-
-type prometheusManager struct {
-	Protocol string
-	// Contains many more fields not listed in this example.
-}
-
-type prometheusManagerCollector struct {
-	prometheusManager *prometheusManager
-}
-
-type prometheusValue struct {
-	Up         int
-	PluginName string
-	HostName   string
-}
-
-func initMetricsServer(addr, port, protocol string) error {
-	log.Info().Str("module", "main").Msgf("start metric server: %s:%s", addr, port)
-
-	reg := prometheus.NewPedanticRegistry()
-
-	var prometheusOnce sync.Once
-
-	prometheusOnce.Do(func() {
-		newPrometheusManager(protocol, reg)
-	})
-
-	reg.MustRegister(
-		prometheus.NewProcessCollector(prometheus.ProcessCollectorOpts{}),
-	)
-
-	http.Handle("/metrics", promhttp.HandlerFor(reg, promhttp.HandlerOpts{}))
-
-	err := http.ListenAndServe(addr+":"+port, nil)
-
-	if err != nil {
-		log.Error().Str("module", "main").Msgf("Prometheus Error: %s", err)
-	}
-
-	return nil
-}
-
-func newPrometheusManager(protocol string, reg prometheus.Registerer) *prometheusManager {
-	c := &prometheusManager{
-		Protocol: protocol,
-	}
-	cc := prometheusManagerCollector{prometheusManager: c}
-	prometheus.WrapRegistererWith(prometheus.Labels{"protocol": protocol}, reg).MustRegister(cc)
-	return c
-}
-
-func (cc prometheusManagerCollector) Describe(ch chan<- *prometheus.Desc) {
-	prometheus.DescribeByCollect(cc, ch)
-}
-
-func (cc prometheusManagerCollector) Collect(ch chan<- prometheus.Metric) {
-	var (
-		pluginUpDesc = prometheus.NewDesc(
-			"plugin_up",
-			"Plugin liveness checks.",
-			[]string{"plugin", "port", "host_name"}, nil,
-		)
-	)
-
-	upByPlugin := cc.prometheusManager.getPluginUp(config.GetConfig().PluginInfos.Plugins, config.GetConfig().Vatz.NotificationInfo.HostName)
-
-	for port, value := range upByPlugin {
-		ch <- prometheus.MustNewConstMetric(
-			pluginUpDesc,
-			prometheus.GaugeValue,
-			float64(value.Up),
-			value.PluginName,
-			strconv.Itoa(port),
-			value.HostName,
-		)
-	}
-}
-
-func (c *prometheusManager) getPluginUp(plugins []config.Plugin, hostName string) (
-	pluginUp map[int]*prometheusValue,
-) {
-	gClients := getClients(plugins)
-	pluginUp = make(map[int]*prometheusValue)
-
-	for idx, plugin := range plugins {
-		pluginUp[plugin.Port] = &prometheusValue{
-			Up:         1,
-			PluginName: plugin.Name,
-			HostName:   hostName,
-		}
-		verify, err := gClients[idx].Verify(context.Background(), new(emptypb.Empty))
-		if err != nil || verify == nil {
-			pluginUp[plugin.Port].Up = 0
-		}
-	}
-
-	return
 }
