+++ utils/utils.go
@@ -1,9 +1,40 @@
 package utils
 
 import (
+	"fmt"
+	pluginpb "github.com/dsrvlabs/vatz-proto/plugin/v1"
+	"github.com/dsrvlabs/vatz/manager/config"
+	"github.com/rs/zerolog/log"
+	"google.golang.org/grpc"
 	"strconv"
 )
 
 func MakeUniqueValue(pName, pAddr string, pPort int) string {
 	return pName + pAddr + strconv.Itoa(pPort)
+}
+
+func GetClients(plugins []config.Plugin) []pluginpb.PluginClient {
+	var grpcClients []pluginpb.PluginClient
+
+	if len(plugins) > 0 {
+		for _, plugin := range plugins {
+			conn, err := grpc.Dial(fmt.Sprintf("%s:%d", plugin.Address, plugin.Port), grpc.WithInsecure())
+			if err != nil {
+				log.Fatal().Str("module", "main").Msgf("gRPC Dial Error(%s): %s", plugin.Name, err)
+			}
+			grpcClients = append(grpcClients, pluginpb.NewPluginClient(conn))
+		}
+	} else {
+		// TODO: Is this really neccessary???
+		defaultConnectedTarget := "localhost:9091"
+		conn, err := grpc.Dial(defaultConnectedTarget, grpc.WithInsecure())
+		if err != nil {
+			log.Fatal().Str("module", "main").Msgf("gRPC Dial Error: %s", err)
+		}
+
+		//TODO: Please, Create a better client functions with static
+		grpcClients = append(grpcClients, pluginpb.NewPluginClient(conn))
+	}
+
+	return grpcClients
 }
